+++ webview-ui/src/components/mcp/configuration/tabs/installed/InstalledServersView.tsx
@@ -1,0 +1,87 @@
+import { VSCodeButton, VSCodeLink } from "@vscode/webview-ui-toolkit/react"
+import { vscode } from "../../../../../utils/vscode"
+import { useExtensionState } from "../../../../../context/ExtensionStateContext"
+import ServerRow from "./server-row/ServerRow"
+const InstalledServersView = () => {
+	const { mcpServers: servers } = useExtensionState()
+
+	return (
+		<div style={{ padding: "16px 20px" }}>
+			<div
+				style={{
+					color: "var(--vscode-foreground)",
+					fontSize: "13px",
+					marginBottom: "16px",
+					marginTop: "5px",
+				}}>
+				The{" "}
+				<VSCodeLink href="https://github.com/modelcontextprotocol" style={{ display: "inline" }}>
+					Model Context Protocol
+				</VSCodeLink>{" "}
+				enables communication with locally running MCP servers that provide additional tools and resources to extend
+				Cline's capabilities. You can use{" "}
+				<VSCodeLink href="https://github.com/modelcontextprotocol/servers" style={{ display: "inline" }}>
+					community-made servers
+				</VSCodeLink>{" "}
+				or ask Cline to create new tools specific to your workflow (e.g., "add a tool that gets the latest npm docs").{" "}
+				<VSCodeLink href="https://x.com/sdrzn/status/1867271665086074969" style={{ display: "inline" }}>
+					See a demo here.
+				</VSCodeLink>
+			</div>
+
+			{servers.length > 0 ? (
+				<div
+					style={{
+						display: "flex",
+						flexDirection: "column",
+						gap: "10px",
+					}}>
+					{servers.map((server) => (
+						<ServerRow key={server.name} server={server} />
+					))}
+				</div>
+			) : (
+				<div
+					style={{
+						display: "flex",
+						flexDirection: "column",
+						alignItems: "center",
+						gap: "12px",
+						marginTop: 20,
+						marginBottom: 20,
+						color: "var(--vscode-descriptionForeground)",
+					}}>
+					No MCP servers installed
+				</div>
+			)}
+
+			{/* Settings Section */}
+			<div style={{ marginBottom: "20px", marginTop: 10 }}>
+				<VSCodeButton
+					appearance="secondary"
+					style={{ width: "100%", marginBottom: "5px" }}
+					onClick={() => {
+						vscode.postMessage({ type: "openMcpSettings" })
+					}}>
+					<span className="codicon codicon-server" style={{ marginRight: "6px" }}></span>
+					Configure MCP Servers
+				</VSCodeButton>
+
+				<div style={{ textAlign: "center" }}>
+					<VSCodeLink
+						onClick={() => {
+							vscode.postMessage({
+								type: "openExtensionSettings",
+								text: "cline.mcp",
+							})
+						}}
+						style={{ fontSize: "12px" }}>
+						Advanced MCP Settings
+					</VSCodeLink>
+				</div>
+			</div>
+		</div>
+	)
+}
+
+export default InstalledServersView
