+++ webview-ui/src/components/mcp/configuration/tabs/installed/server-row/ServerRow.tsx
@@ -1,241 +1,22 @@
+import { McpServer } from "../../../../../../../../src/shared/mcp"
+import { DEFAULT_MCP_TIMEOUT_SECONDS } from "../../../../../../../../src/shared/mcp"
+import { useState } from "react"
+import { vscode } from "../../../../../../utils/vscode"
 import {
 	VSCodeButton,
-	VSCodeLink,
+	VSCodeCheckbox,
+	VSCodeDropdown,
+	VSCodeOption,
 	VSCodePanels,
 	VSCodePanelTab,
 	VSCodePanelView,
-	VSCodeDropdown,
-	VSCodeOption,
-	VSCodeCheckbox,
 } from "@vscode/webview-ui-toolkit/react"
-import { useEffect, useState } from "react"
-import styled from "styled-components"
-import { DEFAULT_MCP_TIMEOUT_SECONDS, McpServer } from "../../../../src/shared/mcp"
-import { useExtensionState } from "../../context/ExtensionStateContext"
-import { getMcpServerDisplayName } from "../../utils/mcp"
-import { vscode } from "../../utils/vscode"
-import McpMarketplaceView from "./marketplace/McpMarketplaceView"
-import McpResourceRow from "./McpResourceRow"
+import { getMcpServerDisplayName } from "../../../../../../utils/mcp"
+import DangerButton from "../../../../../common/DangerButton"
 import McpToolRow from "./McpToolRow"
-import DangerButton from "../common/DangerButton"
-import AddRemoteServerForm from "./tabs/AddRemoteServerForm"
-import AddLocalServerForm from "./tabs/AddLocalServerForm"
+import McpResourceRow from "./McpResourceRow"
+import { useExtensionState } from "../../../../../../context/ExtensionStateContext"
 
-type McpViewProps = {
-	onDone: () => void
-}
-
-const McpView = ({ onDone }: McpViewProps) => {
-	const { mcpServers: servers, mcpMarketplaceEnabled } = useExtensionState()
-	const [activeTab, setActiveTab] = useState(mcpMarketplaceEnabled ? "marketplace" : "installed")
-	const [activeSubTab, setActiveSubTab] = useState("remote")
-
-	const handleTabChange = (tab: string) => {
-		setActiveTab(tab)
-	}
-
-	useEffect(() => {
-		if (!mcpMarketplaceEnabled && activeTab === "marketplace") {
-			// If marketplace is disabled and we're on marketplace tab, switch to installed
-			setActiveTab("installed")
-		}
-	}, [mcpMarketplaceEnabled, activeTab])
-
-	useEffect(() => {
-		if (mcpMarketplaceEnabled) {
-			vscode.postMessage({ type: "silentlyRefreshMcpMarketplace" })
-			vscode.postMessage({ type: "fetchLatestMcpServersFromHub" })
-		}
-	}, [mcpMarketplaceEnabled])
-
-	return (
-		<div
-			style={{
-				position: "fixed",
-				top: 0,
-				left: 0,
-				right: 0,
-				bottom: 0,
-				display: "flex",
-				flexDirection: "column",
-			}}>
-			<div
-				style={{
-					display: "flex",
-					justifyContent: "space-between",
-					alignItems: "center",
-					padding: "10px 17px 5px 20px",
-				}}>
-				<h3 style={{ color: "var(--vscode-foreground)", margin: 0 }}>MCP Servers</h3>
-				<VSCodeButton onClick={onDone}>Done</VSCodeButton>
-			</div>
-
-			<div style={{ flex: 1, overflow: "auto" }}>
-				{/* Tabs container */}
-				<div
-					style={{
-						display: "flex",
-						gap: "1px",
-						padding: "0 20px 0 20px",
-						borderBottom: "1px solid var(--vscode-panel-border)",
-					}}>
-					{mcpMarketplaceEnabled && (
-						<TabButton isActive={activeTab === "marketplace"} onClick={() => handleTabChange("marketplace")}>
-							Marketplace
-						</TabButton>
-					)}
-					<TabButton isActive={activeTab === "addRemote"} onClick={() => handleTabChange("addRemote")}>
-						Add Server
-					</TabButton>
-					<TabButton isActive={activeTab === "installed"} onClick={() => handleTabChange("installed")}>
-						Installed
-					</TabButton>
-				</div>
-
-				{/* Content container */}
-				<div style={{ width: "100%" }}>
-					{mcpMarketplaceEnabled && activeTab === "marketplace" && <McpMarketplaceView />}
-					{activeTab === "addRemote" && (
-						<div style={{ padding: "5px 0 0 0" }}>
-							<div
-								style={{
-									display: "flex",
-									gap: "1px",
-									padding: "0 20px 0 20px",
-									borderBottom: "1px solid var(--vscode-panel-border)",
-								}}>
-								<TabButton isActive={activeSubTab === "remote"} onClick={() => setActiveSubTab("remote")}>
-									Remote (SSE)
-								</TabButton>
-								<TabButton isActive={activeSubTab === "local"} onClick={() => setActiveSubTab("local")}>
-									Local (Command)
-								</TabButton>
-							</div>
-
-							<div>
-								{activeSubTab === "remote" && (
-									<AddRemoteServerForm onServerAdded={() => handleTabChange("installed")} />
-								)}
-								{activeSubTab === "local" && (
-									<AddLocalServerForm onServerAdded={() => handleTabChange("installed")} />
-								)}
-							</div>
-						</div>
-					)}
-					{activeTab === "installed" && (
-						<div style={{ padding: "16px 20px" }}>
-							<div
-								style={{
-									color: "var(--vscode-foreground)",
-									fontSize: "13px",
-									marginBottom: "16px",
-									marginTop: "5px",
-								}}>
-								The{" "}
-								<VSCodeLink href="https://github.com/modelcontextprotocol" style={{ display: "inline" }}>
-									Model Context Protocol
-								</VSCodeLink>{" "}
-								enables communication with locally running MCP servers that provide additional tools and resources
-								to extend Cline's capabilities. You can use{" "}
-								<VSCodeLink href="https://github.com/modelcontextprotocol/servers" style={{ display: "inline" }}>
-									community-made servers
-								</VSCodeLink>{" "}
-								or ask Cline to create new tools specific to your workflow (e.g., "add a tool that gets the latest
-								npm docs").{" "}
-								<VSCodeLink href="https://x.com/sdrzn/status/1867271665086074969" style={{ display: "inline" }}>
-									See a demo here.
-								</VSCodeLink>
-							</div>
-
-							{servers.length > 0 ? (
-								<div
-									style={{
-										display: "flex",
-										flexDirection: "column",
-										gap: "10px",
-									}}>
-									{servers.map((server) => (
-										<ServerRow key={server.name} server={server} />
-									))}
-								</div>
-							) : (
-								<div
-									style={{
-										display: "flex",
-										flexDirection: "column",
-										alignItems: "center",
-										gap: "12px",
-										marginTop: 20,
-										marginBottom: 20,
-										color: "var(--vscode-descriptionForeground)",
-									}}>
-									No MCP servers installed
-								</div>
-							)}
-
-							{/* Settings Section */}
-							<div style={{ marginBottom: "20px", marginTop: 10 }}>
-								<VSCodeButton
-									appearance="secondary"
-									style={{ width: "100%", marginBottom: "5px" }}
-									onClick={() => {
-										vscode.postMessage({ type: "openMcpSettings" })
-									}}>
-									Edit Configuration
-								</VSCodeButton>
-
-								<div style={{ textAlign: "center" }}>
-									<VSCodeLink
-										onClick={() => {
-											vscode.postMessage({
-												type: "openExtensionSettings",
-												text: "cline.mcp",
-											})
-										}}
-										style={{ fontSize: "12px" }}>
-										Advanced MCP Settings
-									</VSCodeLink>
-								</div>
-							</div>
-						</div>
-					)}
-				</div>
-			</div>
-		</div>
-	)
-}
-
-const StyledTabButton = styled.button<{ isActive: boolean }>`
-	background: none;
-	border: none;
-	border-bottom: 2px solid ${(props) => (props.isActive ? "var(--vscode-foreground)" : "transparent")};
-	color: ${(props) => (props.isActive ? "var(--vscode-foreground)" : "var(--vscode-descriptionForeground)")};
-	padding: 8px 16px;
-	cursor: pointer;
-	font-size: 13px;
-	margin-bottom: -1px;
-	font-family: inherit;
-
-	&:hover {
-		color: var(--vscode-foreground);
-	}
-`
-
-export const TabButton = ({
-	children,
-	isActive,
-	onClick,
-}: {
-	children: React.ReactNode
-	isActive: boolean
-	onClick: () => void
-}) => (
-	<StyledTabButton isActive={isActive} onClick={onClick}>
-		{children}
-	</StyledTabButton>
-)
-
-// Server Row Component
 const ServerRow = ({ server }: { server: McpServer }) => {
 	const { mcpMarketplaceCatalog, autoApprovalSettings } = useExtensionState()
 
@@ -576,4 +357,4 @@ const ServerRow = ({ server }: { server: McpServer }) => {
 	)
 }
 
-export default McpView
+export default ServerRow
